Integrating the Ltac language and the Blast tool into the interface capabilities:
The Ltac language is the language that makes it possible to define new tactics
without using the ocaml language (already present in coq for a few months).

The Blast tool is a tool that checks whether the goals could be solve
automatically and proposes the proof trace to the user.


git-svn-id: svn://scm.gforge.inria.fr/svn/coq/trunk@2313 85f007b7-540e-0410-9357-904b9bb8a0f7