New file FMapFullAVL containing the balancing proofs about FMapAVL:
As for FSetAVL vs. FSetFullAVL, preservation of the balancing of trees
is not necessary anymore for just fulfilling the Map interface. But we
still want theses proofs to exists somewhere, since they ensure the
correct efficiency of the FMapAVL operations.

In addition, FSetFullAVL also contains the non-structural,
ocaml-faithful version of compare.



git-svn-id: svn://scm.gforge.inria.fr/svn/coq/trunk@10748 85f007b7-540e-0410-9357-904b9bb8a0f7